Moving the <result> node object from the TestRun.cs file into a TestResult.cs file within TRX for readability
